Add --size flag to resize qcow files on cloning.
Added a flag that allows to resize qcow files on cloning them. When
used after a --name flag it will only apply to the instance group
prior to the --size flag regardless of whether the disk used by that
instance group is global or specific. The same goes for a global
--size flag and instance group specific disks.